+++ section_10/springsecsection_10/src/main/java/com/eazybytes/model/Cards.java
@@ -13,10 +13,10 @@ public class Cards {
 
     @Id
     @Column(name = "card_id")
-    private int cardId;
+    private long cardId;
 
     @Column(name = "customer_id")
-    private int customerId;
+    private long customerId;
 
     @Column(name = "card_number")
     private String cardNumber;
