+++ section_10/springsecsection_10/src/main/java/com/eazybytes/repository/AccountsRepository.java
@@ -1,13 +1,12 @@
 package com.eazybytes.repository;
 
+import com.eazybytes.model.Accounts;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.eazybytes.model.Accounts;
-
 @Repository
-public interface AccountsRepository extends CrudRepository<Accounts, Integer> {
-	
-	Accounts findByCustomerId(int customerId);
+public interface AccountsRepository extends CrudRepository<Accounts, Long> {
+
+    Accounts findByCustomerId(long customerId);
 
 }
