+++ section_10/springsecsection_10/src/main/java/com/eazybytes/repository/LoanRepository.java
@@ -3,15 +3,13 @@ package com.eazybytes.repository;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Repository;
 
 import com.eazybytes.model.Loans;
 
 @Repository
-public interface LoanRepository extends CrudRepository<Loans, Integer> {
-
-	// @PreAuthorize("hasRole('USER')")
-	List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+public interface LoanRepository extends CrudRepository<Loans, Long> {
+	
+	List<Loans> findByCustomerIdOrderByStartDtDesc(long customerId);
 
 }
